fix: unify model metadata lookup sources
