+++ src/providers/hybridai-models.ts
@@ -1,118 +1,9 @@
+import { matchesModelFamily } from './model-lookup.js';
+import { resolveStaticModelCatalogMetadata } from './model-metadata.js';
+
 interface HybridAIModel {
   id: string;
   contextWindowTokens: number | null;
-}
-
-const STATIC_VISION_CAPABLE_MODELS = new Set<string>([
-  'gpt-4.1-mini',
-  'gpt-5',
-  'gpt-5-mini',
-  'gpt-5-pro',
-  'gpt-5.1',
-  'gpt-5.1-codex',
-  'gpt-5.1-codex-max',
-  'gpt-5.1-codex-mini',
-  'gpt-5.2',
-  'gpt-5.2-codex',
-  'gpt-5.2-pro',
-  'gpt-5.4-mini',
-  'gpt-5.3-codex',
-  'gpt-5.4',
-
-  'claude-opus-4-6',
-  'claude-opus-4.6',
-  'claude-sonnet-4-6',
-  'claude-sonnet-4.6',
-
-  'gemini-3',
-  'gemini-3-pro',
-  'gemini-3-flash',
-  'gemini-3.1',
-  'gemini-3.1-pro',
-  'gemini-3.1-pro-high',
-  'gemini-3.1-pro-low',
-  'gemini-3-pro-preview',
-  'gemini-3-flash-preview',
-  'gemini-3.1-pro-preview',
-]);
-
-const STATIC_MODEL_CONTEXT_WINDOWS: Record<string, number> = {
-  'claude-opus-4-6': 200_000,
-  'claude-opus-4.6': 200_000,
-  'claude-sonnet-4-6': 200_000,
-  'claude-sonnet-4.6': 200_000,
-
-  'gemini-3': 1_048_576,
-  'gemini-3-pro': 1_048_576,
-  'gemini-3-flash': 1_048_576,
-  'gemini-3.1': 1_048_576,
-  'gemini-3.1-pro': 1_048_576,
-  'gemini-3.1-pro-high': 1_048_576,
-  'gemini-3.1-pro-low': 1_048_576,
-  'gemini-3-pro-preview': 1_048_576,
-  'gemini-3-flash-preview': 1_048_576,
-  'gemini-3.1-pro-preview': 1_048_576,
-
-  'gpt-4.1-mini': 1_047_576,
-  'gpt-5': 400_000,
-  'gpt-5-chat-latest': 128_000,
-  'gpt-5-codex': 400_000,
-  'gpt-5-mini': 400_000,
-  'gpt-5-nano': 400_000,
-  'gpt-5-pro': 400_000,
-  'gpt-5.1': 400_000,
-  'gpt-5.1-chat-latest': 128_000,
-  'gpt-5.1-codex': 400_000,
-  'gpt-5.1-codex-max': 400_000,
-  'gpt-5.1-codex-mini': 400_000,
-  'gpt-5.2': 400_000,
-  'gpt-5.2-chat-latest': 128_000,
-  'gpt-5.2-codex': 400_000,
-  'gpt-5.2-pro': 400_000,
-  'gpt-5.4-mini': 272_000,
-  'gpt-5.4': 400_000,
-  'gpt-5.3-codex': 400_000,
-  'gpt-5.3-codex-spark': 128_000,
-};
-
-function collectModelLookupCandidates(modelName: string): string[] {
-  const normalized = modelName.trim().toLowerCase();
-  if (!normalized) return [];
-
-  const candidates: string[] = [];
-  const seen = new Set<string>();
-  const queue = [normalized];
-
-  while (queue.length > 0) {
-    const candidate = queue.shift()?.trim().toLowerCase() ?? '';
-    if (!candidate || seen.has(candidate)) continue;
-
-    candidates.push(candidate);
-    seen.add(candidate);
-
-    if (candidate.includes('/')) {
-      queue.push(candidate.split('/').at(-1) ?? '');
-    }
-
-    if (candidate.includes(':')) {
-      queue.push(...candidate.split(':'));
-    }
-  }
-
-  return candidates;
-}
-
-function matchesModelFamily(candidateId: string, targetId: string): boolean {
-  if (!candidateId || !targetId) return false;
-  if (candidateId === targetId) return true;
-  const boundary = candidateId.at(targetId.length);
-  return (
-    candidateId.startsWith(targetId) &&
-    (boundary === '-' ||
-      boundary === '.' ||
-      boundary === ':' ||
-      boundary === '/')
-  );
 }
 
 export function resolveModelContextWindowFromList(
@@ -164,25 +55,7 @@ export function resolveModelContextWindowFromList(
 export function resolveModelContextWindowFallback(
   modelName: string,
 ): number | null {
-  const candidates = collectModelLookupCandidates(modelName);
-  if (candidates.length === 0) return null;
-
-  for (const candidate of candidates) {
-    const direct = STATIC_MODEL_CONTEXT_WINDOWS[candidate];
-    if (direct != null) return direct;
-  }
-
-  // Family fallback for derived ids such as "gpt-5.1-2025-11-13" or
-  // provider/tag forms like "openai/gpt-5:latest".
-  for (const candidate of candidates) {
-    const bestMatch = Object.keys(STATIC_MODEL_CONTEXT_WINDOWS)
-      .filter((key) => matchesModelFamily(candidate, key))
-      .sort((a, b) => b.length - a.length)
-      .at(0);
-    if (bestMatch) return STATIC_MODEL_CONTEXT_WINDOWS[bestMatch] ?? null;
-  }
-
-  return null;
+  return resolveStaticModelCatalogMetadata(modelName).contextWindow;
 }
 
 /**
@@ -192,7 +65,5 @@ export function resolveModelContextWindowFallback(
  * "gpt-5:latest" still match.
  */
 export function isStaticModelVisionCapable(modelName: string): boolean {
-  return collectModelLookupCandidates(modelName).some((candidate) =>
-    STATIC_VISION_CAPABLE_MODELS.has(candidate),
-  );
+  return resolveStaticModelCatalogMetadata(modelName).capabilities.vision;
 }
